converted GT-11141 to enyo 2.6
