+++ test/moonstone/TimePicker/GT-11141-DoesNotLoop-specs.js
@@ -2,7 +2,7 @@ var helpers = rootRequire('./helpers'),
 	app = {};	// Test-specific settings at bottom of the file
 
 var base = 'http://localhost:3000/',
-	url = 'ui-tests/test/loader.html?moonstone/TimePicker/GT-11141-DoesNotLoop',
+	path = 'test/moonstone/TimePicker/GT-11141-DoesNotLoop',
 	title = 'Time Picker: Does Not Loop',
 	tags = ['moonstone', 'TimePicker', 'qa'];	// Tags show up in SauceLabs test output
 
@@ -10,6 +10,7 @@ describe(title, function() {
 	var browser;
 
 	before(function(done) {
+		helpers.epack(path);
 		browser = helpers.initBrowser(title, tags, base, done);
 	});
 
@@ -22,29 +23,29 @@ describe(title, function() {
 	it('Should not loop hours while the minute picker loops', function (done) {
 		browser
 			.setWindowSize(1920,1280)
-			.get(url)
-			.waitForElementById(app.timePickerID)
-			// March 8, 2015 was the date of the daylight savings time change
-			// DST is mentioned in the GT-11141 test details
-			.execute('enyo.$["app"].set("value", new Date("Mar 08 2015 01:59"));')
+			.get("ui-tests/dist")
+			.waitForElementById(app.timePickerID)			
 			.elementById(app.timePickerID)
 				.click()
 			.waitForElementById(app.hourUpArrowID, helpers.wd.asserters.isDisplayed, 1000)
 				.click()
-			.execute('return enyo.$["app"].get("value").getHours()').should.eventually.equal(3)
+			// .enyoPropertyGet(app.appId, "value").getHours().should.eventually.equal(3)
+			.execute('dispatcher = require("enyo/dispatcher"); return dispatcher.$["'+app.appId+'"].get("value").getHours()').should.eventually.equal(3)
 			.elementById(app.minuteUpArrowID)
 				.click()
 				.click()
 				.click()
-			.execute('return enyo.$["app"].get("value").getHours()').should.eventually.equal(3)
+			// .enyoPropertyGet(app.appId, "value").getHours().should.eventually.equal(3)
+			.execute('dispatcher = require("enyo/dispatcher"); return dispatcher.$["'+app.appId+'"].get("value").getHours()').should.eventually.equal(3)
 			.nodeify(done);
 	});
 
 });
 
 app = {
-	timePickerID: 'app_pickerTimeLinked',
-	hourUpArrowID: 'app_pickerTimeLinked_hour_nextOverlay',
-	minuteUpArrowID: 'app_pickerTimeLinked_minute_nextOverlay'
+	appId: "gT-11141-DoesNotLoop",
+	timePickerID: 'gT-11141-DoesNotLoop_pickerTimeLinked',
+	hourUpArrowID: 'gT-11141-DoesNotLoop_pickerTimeLinked_hour_nextOverlay',
+	minuteUpArrowID: 'gT-11141-DoesNotLoop_pickerTimeLinked_minute_nextOverlay'
 };
