+++ test/moonstone/TimePicker/GT-11141-DoesNotLoop/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "GT-11141-DoesNotLoop.js"
+}
